requirements.txt
facenet_pytorch==2.5.2
fastapi==0.79.0
firebase_admin==5.0.3
imutils==0.5.4
Keras==2.1.6
numpy==1.18.5
opencv_contrib_python==4.5.5.64
Pillow==8.4.0
starlette==0.19.1
tensorflow==1.5.0
# torch==1.12.0
uvicorn==0.15.0
anyio>=2.2.0
h5py==2.10.0

# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
# numpy>=1.18.5
# opencv-python>=4.1.1
# Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy==1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
protobuf<3.20,>=3.19.0  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard>=1.5.0
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export onnxsim
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.4.1  # ONNX simplifier
# nvidia-pyindex  # TensorRT export
# nvidia-tensorrt  # TensorRT export
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop>=0.1.1  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow